+++ src/components/form/form_validation_controller.js
@@ -1,0 +1,21 @@
+export default () => ({
+  // cacheDOM() {
+  //   console.log(this.form);
+  //   this.bindEvents();
+  // },
+  // bindEvents() {
+  //   console.log(this.form);
+  // },
+  resetForm() {
+
+  },
+  submitForm() {
+
+  },
+});
+
+// export default function validateForm(formElement) {
+//   // const form = formElement;
+//   // console.log(form);
+//   formValidation.cacheDOM(formElement);
+// }