+++ src/components/print.js
@@ -1,4 +1,0 @@
-// Delete or rename as needed
-export default function printMe() {
-  console.log('I get called from print.js');
-}