+++ src/index.js
@@ -1,18 +1,7 @@
 // Delete or rename as needed
 import './index.css';
-import printMe from './components/print';
+import buildForm from './components/form/form_controller';
 
-function component() {
-  const element = document.createElement('p');
-  element.textContent = 'Hello world';
-
-  const btn = document.createElement('button');
-  btn.innerHTML = 'Click me and check the console!';
-  btn.onclick = printMe;
-
-  element.appendChild(btn);
-
-  return element;
-}
-
-document.body.appendChild(component());
+// const staticForm = buildForm.add('static');
+// document.body.appendChild(staticForm.render());
+document.body.appendChild(buildForm.add('static').render());