+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 ---
+### 24 OCT 2023
+- Created a `form_controller.js` module to control and initialize forms based on a state, created a `form.config.js` module that holds objects to be rendered in the `form.js` module, and initialized `form_validation_controller.js` module.  
+---
+### 23 OCT 2023
+- Initialized module-form, created utility module `createElement.js` that creates an element and assigns `setAttributes()` method to the element, initialized `form.js` and renders a basic form.   
+---
 ### 11 OCT 2023
 - Initial module_webpack_starter package created.  
 ---